backend/requirements.txt
alembic==1.18.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
billiard==4.2.4
black==26.1.0
bs4==0.0.2
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
debugpy==1.8.19
fastapi==0.115.0
flake8==7.3.0
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
isort==7.0.0
jinja2==3.1.6
kombu==5.6.2
mako==1.3.10
markupsafe==3.0.3
mccabe==0.7.0
mypy-extensions==1.1.0
packaging==26.0
pathspec==1.0.3
platformdirs==4.5.1
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pydantic==2.9.2
pydantic-core==2.23.4
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytokens==0.4.0
pyyaml==6.0.3
redis==7.1.0
requests==2.32.5
ruff==0.14.14
six==1.17.0
soupsieve==2.8.3
sqlalchemy==2.0.46
sqlmodel==0.0.31
starlette==0.38.6
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.32.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.3.3
websockets==16.0
